# v2: errors